Remove legacy entity structures in favor of Diesel-based models and migrations
Deleted outdated entity structs, views, and related code for building data access layers. Consolidated logic into the new Diesel-based system with migrations for schema management. Enhanced modularity and code maintainability.

+++ lib-tauri/app_models/migrations/postgres/2025-07-02-074045_create_dati_stanza_view/up.sql
@@ -43,7 +43,7 @@ select s.id,
        s.riscaldamento,
        s.raffrescamento,
        s.illuminazione,
-       coalesce(round(dgs.mq_infissi::numeric, 2), 0) as mq_infissi,
+       coalesce(round(dgs.mq_infissi::numeric, 2), 0)::float as mq_infissi,
        dgs.materiale,
        dgs.vetro
 from stanza as s
